JudoOptions moved to correct package
