+++ judo-sdk/src/main/java/com/judopay/JudoOptions.java
@@ -1,4 +1,4 @@
-package com.judopay.payment.form;
+package com.judopay;
 
 import android.os.Bundle;
 import android.os.Parcel;
