+++ judo-sdk/src/main/java/com/judopay/PaymentActivity.java
@@ -1,13 +1,10 @@
 package com.judopay;
 
 import android.os.Bundle;
-import android.os.Parcelable;
 
 import com.judopay.model.Receipt;
-import com.judopay.payment.form.JudoOptions;
 
 import static com.judopay.Judo.JUDO_AMOUNT;
-import static com.judopay.Judo.JUDO_CARD_TOKEN;
 import static com.judopay.Judo.JUDO_CONSUMER;
 import static com.judopay.Judo.JUDO_CURRENCY;
 import static com.judopay.Judo.JUDO_ID;
