+++ judo-sdk/src/main/java/com/judopay/RegisterCardActivity.java
@@ -2,12 +2,7 @@ package com.judopay;
 
 import android.os.Bundle;
 
-import com.judopay.payment.form.JudoOptions;
-
-import static com.judopay.Judo.JUDO_AMOUNT;
-import static com.judopay.Judo.JUDO_CARD_TOKEN;
 import static com.judopay.Judo.JUDO_CONSUMER;
-import static com.judopay.Judo.JUDO_CURRENCY;
 import static com.judopay.Judo.JUDO_ID;
 import static com.judopay.Judo.JUDO_OPTIONS;
