+++ judo-sdk/src/main/java/com/judopay/RegisterCardPresenter.java
@@ -3,7 +3,6 @@ package com.judopay;
 import com.google.gson.Gson;
 import com.judopay.model.Card;
 import com.judopay.model.RegisterTransaction;
-import com.judopay.payment.form.JudoOptions;
 
 class RegisterCardPresenter extends BasePresenter {
