+++ judo-sdk/src/main/java/com/judopay/TokenPaymentActivity.java
@@ -2,7 +2,6 @@ package com.judopay;
 
 import android.os.Bundle;
 
-import com.judopay.payment.form.JudoOptions;
 import static com.judopay.Judo.JUDO_AMOUNT;
 import static com.judopay.Judo.JUDO_CARD_TOKEN;
 import static com.judopay.Judo.JUDO_CONSUMER;
