+++ judo-sdk/src/main/java/com/judopay/TokenPreAuthFragment.java
@@ -6,7 +6,6 @@ import android.view.View;
 import com.google.gson.Gson;
 import com.judopay.api.JudoApiServiceFactory;
 import com.judopay.model.Card;
-import com.judopay.payment.form.JudoOptions;
 
 public class TokenPreAuthFragment extends BaseFragment {
