--- judo-sdk/src/main/java/com/judopay/payment/form/PaymentFormListener.java
@@ -1,9 +1,0 @@
-package com.judopay.payment.form;
-
-import com.judopay.model.Card;
-
-public interface PaymentFormListener {
-
-    void onSubmit(Card card);
-
-}
