+++ judo-sdk/src/test/java/com/judopay/TokenPreAuthPresenterTest.java
@@ -7,7 +7,6 @@ import com.judopay.model.CardToken;
 import com.judopay.model.Currency;
 import com.judopay.model.Receipt;
 import com.judopay.model.TokenTransaction;
-import com.judopay.payment.form.JudoOptions;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
